backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email-validator]
alembic
python-multipart
opencv-python-headless
numpy
Pillow
rembg
faster-whisper
scenedetect[opencv]
aiofiles
google-generativeai
